% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sport}
\alias{plot_sport}
\title{Plot sport search percentage against Trump voting percentage}
\usage{
plot_sport(sport = "all")
}
\arguments{
\item{sport}{A string, has to be one of the sports in \code{\link{sportsPolitics}}. "all" is the default.}
}
\value{
A scatter plot of the selected sport, if sport = "all" returns a plot for all sports
}
\description{
Plot search percentage against Trump voting percentage in \code{\link{sportsPolitics}}
}
